requirements.txt
# Python
certifi==2022.6.15
charset-normalizer==2.0.12
idna==3.3
python-dotenv==0.20.0
requests==2.28.0
urllib3==1.26.9


# linkedin-api
linkedin-api~=2.0.0a5
beautifulsoup4==4.11.1
lxml==4.9.0
soupsieve==2.3.2.post1


# flask
Jinja2==3.1.2
MarkupSafe==2.1.1
Werkzeug==2.1.2
flask==2.1.2
importlib-metadata==4.12.0
itsdangerous==2.1.2
zipp==3.8.0
gunicorn==20.1.0

# SQLAlchemy
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.39
greenlet==1.1.2
psycopg2-binary==2.9.3

# celery
amqp==5.1.1
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.5
deprecated==1.2.13
kombu==5.2.4
packaging==21.3
prompt-toolkit==3.0.30
pyparsing==3.0.9
pytz==2022.1
redis==4.3.4
six==1.16.0
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
flower==1.1.0
humanize==4.2.3
prometheus-client==0.14.1
tornado==6.2
